# src/IO.Swagger/Controllers/DietApi.cs
/*
 * Training Journal - OpenAPI 3.0
 *
 * This a personal training journal API Server based on the OpenAPI 3.0 specification.  ### GitHub repo of the project: - [My GitHub](https://github.com/aminsys/training-journal)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DietApiController : ControllerBase
    { 
        /// <summary>
        /// Get the user&#x27;s collection of registered diets.
        /// </summary>
        /// <response code="200">successful request</response>
        /// <response code="404">Diet collection not found for the user</response>
        [HttpGet]
        [Route("/api/v1/user/diet")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("UserDietGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(Diet), description: "successful request")]
        public virtual IActionResult UserDietGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Diet));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"Info\" : \"Chest workout\",\n  \"DietId\" : 62\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Diet>(exampleJson)
                        : default(Diet);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}