# src/IO.Swagger/Controllers/TrainingSessionsApi.cs
/*
 * Training Journal - OpenAPI 3.0
 *
 * This a personal training journal API Server based on the OpenAPI 3.0 specification.  ### GitHub repo of the project: - [My GitHub](https://github.com/aminsys/training-journal)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class TrainingSessionsApiController : ControllerBase
    { 
        /// <summary>
        /// Get a user&#x27;s single training session by ID.
        /// </summary>
        /// <param name="trainingSessionId">Training session&#x27;s ID</param>
        /// <response code="200">successful request</response>
        /// <response code="404">Training session is not found</response>
        [HttpGet]
        [Route("/api/v1/user/trainingSessions/{trainingSessionId}")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetTrainingSessionById")]
        [SwaggerResponse(statusCode: 200, type: typeof(TrainingSession), description: "successful request")]
        public virtual IActionResult GetTrainingSessionById([FromRoute][Required]long? trainingSessionId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(TrainingSession));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"exercises\" : [ \"exercises\", \"exercises\" ],\n  \"trainingSessionId\" : 3,\n  \"trainingsName\" : \"Chest and triceps workout\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<TrainingSession>(exampleJson)
                        : default(TrainingSession);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get a user&#x27;s collection of training sessions.
        /// </summary>
        /// <response code="200">successful request</response>
        /// <response code="404">Training sessions not found for the user</response>
        [HttpGet]
        [Route("/api/v1/user/trainingSessions")]
        [Authorize(AuthenticationSchemes = BasicAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("UserTrainingSessionsGet")]
        [SwaggerResponse(statusCode: 200, type: typeof(TrainingSessions), description: "successful request")]
        public virtual IActionResult UserTrainingSessionsGet()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(TrainingSessions));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"trainingSessionId\" : 3,\n  \"trainingsName\" : \"Chest workout\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<TrainingSessions>(exampleJson)
                        : default(TrainingSessions);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}