# src/IO.Swagger/Models/TrainingSession.cs
/*
 * Training Journal - OpenAPI 3.0
 *
 * This a personal training journal API Server based on the OpenAPI 3.0 specification.  ### GitHub repo of the project: - [My GitHub](https://github.com/aminsys/training-journal)
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace IO.Swagger.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class TrainingSession : IEquatable<TrainingSession>
    { 
        /// <summary>
        /// Training session&#x27;s ID
        /// </summary>
        /// <value>Training session&#x27;s ID</value>

        [DataMember(Name="trainingSessionId")]
        public long? TrainingSessionId { get; set; }

        /// <summary>
        /// Name of training session
        /// </summary>
        /// <value>Name of training session</value>

        [DataMember(Name="trainingsName")]
        public string TrainingsName { get; set; }

        /// <summary>
        /// Gets or Sets Exercises
        /// </summary>

        [DataMember(Name="exercises")]
        public List<string> Exercises { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class TrainingSession {\n");
            sb.Append("  TrainingSessionId: ").Append(TrainingSessionId).Append("\n");
            sb.Append("  TrainingsName: ").Append(TrainingsName).Append("\n");
            sb.Append("  Exercises: ").Append(Exercises).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((TrainingSession)obj);
        }

        /// <summary>
        /// Returns true if TrainingSession instances are equal
        /// </summary>
        /// <param name="other">Instance of TrainingSession to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(TrainingSession other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    TrainingSessionId == other.TrainingSessionId ||
                    TrainingSessionId != null &&
                    TrainingSessionId.Equals(other.TrainingSessionId)
                ) && 
                (
                    TrainingsName == other.TrainingsName ||
                    TrainingsName != null &&
                    TrainingsName.Equals(other.TrainingsName)
                ) && 
                (
                    Exercises == other.Exercises ||
                    Exercises != null &&
                    Exercises.SequenceEqual(other.Exercises)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (TrainingSessionId != null)
                    hashCode = hashCode * 59 + TrainingSessionId.GetHashCode();
                    if (TrainingsName != null)
                    hashCode = hashCode * 59 + TrainingsName.GetHashCode();
                    if (Exercises != null)
                    hashCode = hashCode * 59 + Exercises.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(TrainingSession left, TrainingSession right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(TrainingSession left, TrainingSession right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}